# backend/openapi.yaml
openapi: 3.0.3
info:
  title: RavelGo API
  version: "0.1.0"
  description: >
    Unified REST API for the RavelGo rider, driver, and admin apps.
    Authenticated with Amazon Cognito access tokens; role checks are enforced
    via Cognito group membership (`Rider`, `Driver`, `Admin`) — see
    `src/middleware/auth.ts`. There is also a WebSocket endpoint
    (`wss://<host>/ws?token=<access token>`) for live trip-status and
    driver-location push, not representable in OpenAPI 3.0 — see
    `docs/realtime-architecture.md` for its wire protocol. Every REST
    endpoint below has an HTTP-polling equivalent, so a client never
    strictly needs the WebSocket connection.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /riders/me/loyalty:
    get:
      tags: [Loyalty & Promotions]
      summary: My current loyalty tier, computed from completed trips (Rider)
      responses:
        "200":
          description: Completed trip count and computed tier
          content:
            application/json:
              schema:
                type: object
                properties:
                  completedTrips: { type: integer }
                  currentTier:
                    oneOf:
                      - $ref: "#/components/schemas/LoyaltyTier"
                      - nullable: true
        "404": { $ref: "#/components/responses/NotFound" }
...
  /loyalty-tiers:
    get:
      tags: [Loyalty & Promotions]
      summary: Browse the loyalty tier ladder (any authenticated user)
      responses:
        "200":
          description: Tiers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LoyaltyTier" }
    post:
      tags: [Loyalty & Promotions]
      summary: Define a loyalty tier (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, minCompletedTrips, discountPercent]
              properties:
                name: { type: string }
                minCompletedTrips: { type: integer }
                discountPercent: { type: number }
                perks: { type: string }
      responses:
        "201":
          description: Tier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoyaltyTier" }
  /promotions:
    get:
      tags: [Loyalty & Promotions]
      summary: Browse active, unexpired promotions (any authenticated user)
      responses:
        "200":
          description: Promotions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Promotion" }
    post:
      tags: [Loyalty & Promotions]
      summary: Create a promotion code (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, description, discountPercent]
              properties:
                code: { type: string }
                description: { type: string }
                discountPercent: { type: number }
                expiresAt: { type: string, format: date-time }
                maxRedemptions: { type: integer }
      responses:
        "201":
          description: Promotion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Promotion" }
  /promotions/{code}/redeem:
    post:
      tags: [Loyalty & Promotions]
      summary: Redeem a promotion code, once per user (any authenticated user)
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Redemption + discount
          content:
            application/json:
              schema:
                type: object
                properties:
                  redemption: { $ref: "#/components/schemas/PromotionRedemption" }
                  discountPercent: { type: number }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Expired, redemption limit reached, or already redeemed by this user
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Cognito access token
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    LoyaltyTier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        minCompletedTrips: { type: integer }
        discountPercent: { type: number }
        perks: { type: string, nullable: true }
    Promotion:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        description: { type: string }
        discountPercent: { type: number }
        active: { type: boolean }
        expiresAt: { type: string, format: date-time, nullable: true }
        maxRedemptions: { type: integer, nullable: true }
        redemptionCount: { type: integer }
    PromotionRedemption:
      type: object
      properties:
        id: { type: string }
        promotionId: { type: string }
        userId: { type: string }
        redeemedAt: { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_SERVER_ERROR
                - RATE_LIMIT_EXCEEDED
                - BAD_REQUEST
                - PAYMENT_ERROR
                - DATABASE_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              nullable: true
              description: Field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string, description: 'Field name or path (e.g., "email", "address.zipCode")' }
                  message: { type: string, description: Error message for this field }
                  code: { type: string, nullable: true, description: 'Error code (e.g., "too_small", "invalid_email")' }
            timestamp:
              type: string
              format: date-time
              description: When the error occurred (ISO 8601)